+++ common/binary_search_tree/info/update_info_with_path.h
@@ -1,17 +1,19 @@
 #pragma once
 
+#include "common/base.h"
 #include "common/template.h"
+
 #include <vector>
 
 namespace bst {
 template <class TNode>
 inline void UpdateInfoNodeToRootWithPath(
-    const std::vector<TNode*>& node_to_root_path, unsigned start_from_index,
+    const std::vector<TNode*>& node_to_root_path, size_t start_from_index,
     TFakeFalse) {}
 
 template <class TNode>
 inline void UpdateInfoNodeToRootWithPath(
-    const std::vector<TNode*>& node_to_root_path, unsigned start_from_index,
+    const std::vector<TNode*>& node_to_root_path, size_t start_from_index,
     TFakeTrue) {
   for (auto it = node_to_root_path.begin() + start_from_index;
        it != node_to_root_path.end(); ++it)
@@ -20,8 +22,7 @@ inline void UpdateInfoNodeToRootWithPath(
 
 template <class TNode>
 inline void UpdateInfoNodeToRootWithPath(
-    const std::vector<TNode*>& node_to_root_path,
-    unsigned start_from_index = 0) {
+    const std::vector<TNode*>& node_to_root_path, size_t start_from_index = 0) {
   UpdateInfoNodeToRootWithPath(node_to_root_path, start_from_index,
                                TFakeBool<!TNode::TInfo::is_none>());
 }
